+++ src/graph/index.ts
@@ -1,3 +1,4 @@
 export * from './database.js'
 export * from './edges.js'
 export * from './nodes.js'
+export * from './file-ir.js'